Major refactor: Created Job Object interfaces and inject OS-specific implementations with NInject
- Instantiate NonInteractiveProcess, BackgroundProcessWorker, FFmpegPlugin, FFmpeg, and MkvPropEdit with NInject
- Added PInvokeUtils.cs to OSUtils

+++ Libraries/WindowsOSUtils/JobObjects/JobObjectController.cs
@@ -356,7 +356,7 @@ namespace WindowsOSUtils.JobObjects
                   _currentJobHandle,
                   JobObjectInfoClass.ExtendedLimitInformation,
                   ref jobObjectInfo,
-                  Marshal.SizeOf(typeof(ExtendedLimits32)));
+                  (uint) Marshal.SizeOf(typeof(ExtendedLimits32)));
 
                 if (!result)
                 {
@@ -392,7 +392,7 @@ namespace WindowsOSUtils.JobObjects
                   _currentJobHandle,
                   JobObjectInfoClass.ExtendedLimitInformation,
                   ref jobObjectInfo,
-                  Marshal.SizeOf(typeof(ExtendedLimits64)));
+                  (uint) Marshal.SizeOf(typeof(ExtendedLimits64)));
 
                 if (!result)
                 {
